+++ packages/atproto_core/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release Note
 
+## v0.1.0
+
+- Added core objects. ([#70](https://github.com/myConsciousness/atproto.dart/issues/70))
+  - `BaseService`
+  - `Empty`
+  - `ATProtoResponse`
+  - `ATProtoRequest`
+  - `ATProtoException`
+  - `ForbiddenException`
+
 ## v0.0.2
 
 - Exposed `client_context.dart`.